Make changes

+++ src/main/java/pro/sky/java/cours2/webcalculator/service/Service.java
@@ -1,9 +1,9 @@
 package pro.sky.java.cours2.webcalculator.service;
 
 public interface Service {
-    String printPlus(Double a, Double b);
-    String printMinus(Double a, Double b);
-    String printMultiply(Double a, Double b);
-    String printDivide(Double a, Double b);
+    String printPlus(Integer a, Integer b);
+    String printMinus(Integer a, Integer b);
+    String printMultiply(Integer a, Integer b);
+    String printDivide(Integer a, Integer b);
 
 }
